refactor(browser): tighten the scheduler seam after the split
The split left each scheduler rebuilding the controller's closure by hand:
pure helpers arrived as injected callbacks, and both branches assembled the
run's public result themselves, which forced a type import back into the
controller and fossilized a duplicated epilogue.

- Name the seam: `schedulerSeam.ts` owns `BrowserWatchSession`,
  `SchedulerRunResult`, and `DispatchPageResolver`, so neither scheduler
  reaches back into `hostController` for a type.
- Give the watch handover a home: `watchSignals.ts` replaces the anonymous
  box, and both schedulers narrow it with `Pick` instead of restating it —
  the two hand-written copies had already drifted apart in arity.
- Import the closure-free helpers instead of injecting them; the injected
  `mapViewportByProject` signature had drifted from its implementation, and
  the unit test was asserting against four reimplemented fakes rather than
  the production code. `collectDeletedTestPaths` moves to the module that
  owns file-set diffing.
- Assemble the run result once in the controller: schedulers report only
  what they alone know, which drops seven pass-through deps per branch.
- Name the test files after the modules they cover.
