+++ packages/browser/src/watchRerunPlanner.ts
@@ -1,6 +1,21 @@
 import { normalize } from 'pathe';
 import type { TestFileInfo } from './protocol';
 
+/**
+ * Paths the previous cycle ran that the current file set no longer contains.
+ * Core prunes its own state from this, so a file deleted mid-session stops
+ * being reported instead of lingering as a passing result.
+ */
+export const collectDeletedTestPaths = (
+  previous: TestFileInfo[],
+  current: TestFileInfo[],
+): string[] => {
+  const currentPathSet = new Set(current.map((file) => file.testPath));
+  return previous
+    .map((file) => file.testPath)
+    .filter((testPath) => !currentPathSet.has(testPath));
+};
+
 type WatchPlannerProjectEntry = {
   project: {
     name: string;
